--- diagram/diagram.html
@@ -1,64 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-	<meta charset="UTF-8">
-	<meta http-equiv="X-UA-Compatible" content="IE=edge">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<title>Hypercomplex Numbers Diagram</title>
-</head>
-
-<style>
-	body {
-		background-color: #222;
-	}
-
-	div {
-		border: solid 2px #222;
-		padding: 6px;
-		font-size: 18px;
-		color: #0F0F0F;
-		font-family: "Calibri";
-		font-weight: 460;
-		font-weight: 600;
-		border-radius: 1000px;
-		display: inline-block;
-	}
-
-	p {
-		margin: 6px;
-		display: inline;
-	}
-</style>
-
-<body>
-	<div style="background: #6BC872;">
-		<div style="background: #FD6460;">
-			<div style="background: #ADFEF0;">
-				<div style="background: #3FA2E3;">
-					<div style="background: #C5E36F;">
-						<div style="background: #F88553;">
-							<div style="background: #F45E7D;">
-								<div style="background: #E9F65D;">
-									<div style="background: #F4F4F4;">
-										<p>Real</p>
-									</div>
-									<p>Complex</p>
-								</div>
-								<p>Quaternion</p>
-							</div>
-							<p>Octonion</p>
-						</div>
-						<p>Sedenion</p>
-					</div>
-					<p>Pathion</p>
-				</div>
-				<p>Chingon</p>
-			</div>
-			<p>Routon</p>
-		</div>
-		<p>Voudon</p>
-	</div>
-</body>
-
-</html>
