+++ hypercomplex_original.py
