--- tox.ini
@@ -1,8 +1,0 @@
-[tox]
-envlist = py39, py38, py37, py36
-
-[testenv]
-deps =
-
-commands =
-    python -m unittest discover -s ./hypercomplex
